refactor level generator code
